W-14875628 (refactor): extracted async declaration parsing to separate class
