+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/document/AsyncApiDocumentParser.scala
@@ -71,6 +71,16 @@ abstract class AsyncApiDocumentParser(root: Root)(implicit val ctx: AsyncWebApiC
     document
   }
 
+  private def parseDeclarations(map: YMap, parentObj: AmfObject) = {
+    map.key("components").foreach { components =>
+      val parent        = root.location + "#/declarations"
+      val componentsMap = components.value.as[YMap]
+      AsyncApiDeclarationParser(definitionsKey).parseDeclarations(componentsMap, parent)
+      ctx.closedShape(parentObj, componentsMap, "components")
+      validateNames()
+    }
+  }
+
   def parseApi(map: YMap): AsyncApi = {
     YamlTagValidator.validate(root)
     val api = AsyncApi(root.parsed.asInstanceOf[SyamlParsedDocument].document.node)
@@ -134,187 +144,6 @@ abstract class AsyncApiDocumentParser(root: Root)(implicit val ctx: AsyncWebApiC
   }
 
   override protected val definitionsKey: String = "schemas"
-
-  def parseDeclarations(map: YMap, parentObj: AmfObject): Unit = {
-    map.key("components").foreach { components =>
-      val parent        = root.location + "#/declarations"
-      val componentsMap = components.value.as[YMap]
-
-      parseSecuritySchemeDeclarations(componentsMap, parent + "/securitySchemes")
-      parseCorrelationIdDeclarations(componentsMap, parent + "/correlationIds")
-      super.parseTypeDeclarations(componentsMap, Some(this))
-      parseParameterDeclarations(componentsMap, parent + "/parameters")
-
-      parseMessageBindingsDeclarations(componentsMap, parent + "/messageBindings")
-      parseServerBindingsDeclarations(componentsMap, parent + "/serverBindings")
-      parseOperationBindingsDeclarations(componentsMap, parent + "/operationBindings")
-      parseChannelBindingsDeclarations(componentsMap, parent + "/channelBindings")
-      parseOperationTraits(componentsMap, parent + "/operationTraits")
-      parseMessageTraits(componentsMap, parent + "/messageTraits")
-
-      parseMessageDeclarations(componentsMap, parent + "/messages")
-
-      ctx.closedShape(parentObj, componentsMap, "components")
-      validateNames()
-    }
-  }
-
-  def parseMessageDeclarations(componentsMap: YMap, parent: String): Unit =
-    componentsMap.key(
-      "messages",
-      e => {
-        addDeclarationKey(DeclarationKey(e))
-        e.value.as[YMap].entries.foreach { entry =>
-          val message = AsyncMessageParser(YMapEntryLike(entry), parent, None).parse()
-          message.add(DeclaredElement())
-          ctx.declarations += message
-        }
-      }
-    )
-
-  def parseOperationTraits(componentsMap: YMap, parent: String): Unit =
-    componentsMap.key(
-      "operationTraits",
-      entry => {
-        addDeclarationKey(DeclarationKey(entry, isAbstract = true))
-        entry.value.as[YMap].entries.foreach { entry =>
-          val adopt     = (o: Operation) => o
-          val operation = AsyncOperationParser(entry, adopt, isTrait = true).parse()
-          operation.add(DeclaredElement())
-          ctx.declarations += operation
-        }
-      }
-    )
-
-  def parseMessageTraits(componentsMap: YMap, parent: String): Unit =
-    componentsMap.key(
-      "messageTraits",
-      entry => {
-        addDeclarationKey(DeclarationKey(entry, isAbstract = true))
-        entry.value.as[YMap].entries.foreach { entry =>
-          val message = AsyncMessageParser(YMapEntryLike(entry), parent, None, isTrait = true).parse()
-          message.add(DeclaredElement())
-          ctx.declarations += message
-        }
-      }
-    )
-
-  protected def parseSecuritySchemeDeclarations(map: YMap, parent: String): Unit = {
-    map.key(
-      "securitySchemes",
-      e => {
-        addDeclarationKey(DeclarationKey(e))
-        e.value.as[YMap].entries.foreach { entry =>
-          ctx.declarations += ctx.factory
-            .securitySchemeParser(
-              entry,
-              (scheme) => {
-                val name = entry.key.as[String]
-                scheme.setWithoutId(
-                  SecuritySchemeModel.Name,
-                  AmfScalar(name, Annotations(entry.key.value)),
-                  Annotations(entry.key)
-                )
-                scheme
-              }
-            )
-            .parse()
-            .add(DeclaredElement())
-        }
-      }
-    )
-  }
-
-  def parseParameterDeclarations(componentsMap: YMap, parent: String): Unit = {
-    componentsMap.key(
-      "parameters",
-      paramsMap => {
-        addDeclarationKey(DeclarationKey(paramsMap))
-        val parameters: Seq[Parameter] = AsyncParametersParser(parent, paramsMap.value.as[YMap]).parse()
-        parameters map { param =>
-          param.add(DeclaredElement())
-          ctx.declarations += param
-        }
-      }
-    )
-  }
-
-  def parseCorrelationIdDeclarations(componentsMap: YMap, parent: String): Unit = {
-    componentsMap.key(
-      "correlationIds",
-      e => {
-        addDeclarationKey(DeclarationKey(e))
-        e.value.as[YMap].entries.foreach { entry =>
-          val correlationId = AsyncCorrelationIdParser(YMapEntryLike(entry), parent).parse()
-          ctx.declarations += correlationId.add(DeclaredElement())
-        }
-      }
-    )
-  }
-
-  def parseMessageBindingsDeclarations(componentsMap: YMap, parent: String): Unit = {
-    parseBindingsDeclarations[MessageBindings](
-      "messageBindings",
-      componentsMap,
-      entry => {
-        AsyncMessageBindingsParser(YMapEntryLike(entry)).parse()
-      },
-      MessageBindingsModel
-    )
-  }
-
-  def parseServerBindingsDeclarations(componentsMap: YMap, parent: String): Unit = {
-    parseBindingsDeclarations[ServerBindings](
-      "serverBindings",
-      componentsMap,
-      entry => {
-        AsyncServerBindingsParser(YMapEntryLike(entry)).parse()
-      },
-      ServerBindingsModel
-    )
-  }
-
-  def parseOperationBindingsDeclarations(componentsMap: YMap, parent: String): Unit = {
-    parseBindingsDeclarations[OperationBindings](
-      "operationBindings",
-      componentsMap,
-      entry => {
-        AsyncOperationBindingsParser(YMapEntryLike(entry)).parse()
-      },
-      OperationBindingsModel
-    )
-  }
-
-  def parseChannelBindingsDeclarations(componentsMap: YMap, parent: String): Unit = {
-    parseBindingsDeclarations[ChannelBindings](
-      "channelBindings",
-      componentsMap,
-      entry => {
-        AsyncChannelBindingsParser(YMapEntryLike(entry)).parse()
-      },
-      ChannelBindingsModel
-    )
-  }
-
-  def parseBindingsDeclarations[T <: DomainElement](
-      keyword: String,
-      componentsMap: YMap,
-      parse: YMapEntry => T,
-      model: DomainElementModel
-  ): Unit = {
-    componentsMap.key(
-      keyword,
-      e => {
-        addDeclarationKey(DeclarationKey(e))
-        e.value.as[YMap].entries.foreach { entry =>
-          val bindings: T = parse(entry)
-          bindings.add(DeclaredElement())
-          ctx.declarations += bindings
-        }
-      }
-    )
-  }
-
 }
 
 case class IdentifierParser(entry: YMapEntry, webApi: AsyncApi, override implicit val ctx: AsyncWebApiContext)
